+++ crates/rpc-types-eth/src/block.rs
@@ -83,7 +83,7 @@ pub struct Header {
     pub gas_limit: u64,
     /// Gas Used
     #[cfg_attr(feature = "serde", serde(default, with = "alloy_serde::quantity"))]
-    pub gas_used: u128,
+    pub gas_used: u64,
     /// Timestamp
     #[cfg_attr(feature = "serde", serde(default, with = "alloy_serde::quantity"))]
     pub timestamp: u64,
